MitraVerify-Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
transformers==4.45.0
torch==2.6.0
sentence-transformers==3.1.1
Pillow==11.0.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
requests==2.32.3
jinja2==3.1.4
python-dotenv==1.0.1
imagehash==4.3.1
opencv-python==4.10.0.84
matplotlib==3.9.2